ml-service/requirements.txt
lightgbm==4.3.0
onnxruntime==1.17.3
onnxmltools==1.12.0
torch==2.3.0
torch-geometric==2.5.3
aif360==0.6.1
shap==0.44.1
grpcio==1.63.0
grpcio-tools==1.63.0
kafka-python==2.0.2
boto3==1.34.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
prometheus-client==0.20.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-grpc==0.44b0
psycopg2-binary==2.9.9
pytest==8.1.1
